Move fetch project action to new projects store
